chore: bk separate into components
